+++ src/main/kotlin/org/axonframework/intellij/ide/plugin/api/PsiElementWrapper.kt
@@ -64,4 +64,10 @@ interface PsiElementWrapper {
      * @return The correct icon to be used in a line marker popup
      */
     fun getIcon(): Icon
+
+    /**
+     * Retrieve the key to sort the list with. By default takes icon and text into account to show it nicely categorized.
+     * Override this method for customized behavior.
+     */
+    fun getSortKey() = getIcon().toString() + renderText()
 }