+++ src/main/kotlin/org/axonframework/intellij/ide/plugin/resolving/AggregateStructureResolver.kt
@@ -42,32 +42,38 @@ import org.axonframework.intellij.ide.plugin.util.toQualifiedName
 class AggregateStructureResolver(private val project: Project) {
     private val cache = project.createCachedValue { resolve() }
 
-    fun getModels(): List<Model> = cache.value
-
-    fun getFlattendedModelsAndEntities(): List<Model> = getModels().flatMap { it.flatten() }
-
-    fun getAllModelsRelatedToName(name: String): List<Model> {
-        getMemberForName(name) ?: return emptyList()
-        return getModels().firstOrNull { it.contains(name) }?.flatten() ?: emptyList()
+    /**
+     * Finds the uppermost Model owning the Model with this name.
+     * So, If Aggregate A owns entity B, which in turns owns entity C. When requesting the owner for C, it will return A.
+     */
+    fun getHierarchyOwnerForName(name: String): Model? {
+        val member = getMemberForName(name) ?: return null
+        if (member.parent != null) {
+            return getHierarchyOwnerForName(member.parent)
+        }
+        return member
     }
 
-    fun getParentOfModelChild(name: String): List<ModelChild> {
-        return getFlattendedModelsAndEntities()
-            .flatMap { it.children }
-            .filter { c -> c.member.name == name }
-    }
-
-    private fun Model.contains(name: String): Boolean {
-        return this.name == name || children.any { child -> child.member.contains(name) }
-    }
-
+    /**
+     * Returns the Model for this name
+     */
     fun getMemberForName(name: String): Model? {
-        return getFlattendedModelsAndEntities().firstOrNull { it.name == name }
+        return cache.value.firstOrNull { it.name == name }
     }
 
+    /**
+     * Finds the model with given name and all it's sub-entities down in the hierarchy.
+     */
     fun getMemberWithSubEntities(name: String): List<Model> {
         val member = getMemberForName(name) ?: return emptyList()
-        return member.flatten()
+        return getMemberWithSubEntities(member)
+    }
+
+    private fun getMemberWithSubEntities(model: Model): List<Model> {
+        if(model.children.isEmpty()) {
+            return listOf(model)
+        }
+        return model.children.flatMap { getMemberWithSubEntities(it.member) }
     }
 
     private fun Model.flatten() = listOf(this) + children.map { it.member }
@@ -77,9 +83,10 @@ class AggregateStructureResolver(private val project: Project) {
         .flatMap {
             AnnotatedElementsSearch.searchPsiClasses(it.psiClass, project.axonScope()).findAll()
         }
-        .mapNotNull { inspect(it) }
+        .mapNotNull { inspect(it, null) }
+        .flatMap { it.flatten() }
 
-    private fun inspect(clazz: PsiClass): Model? {
+    private fun inspect(clazz: PsiClass, parent: PsiClass?): Model? {
         if (clazz.isEnum) {
             return null
         }
@@ -91,10 +98,10 @@ class AggregateStructureResolver(private val project: Project) {
                 val qualifiedName = psiType.toQualifiedName() ?: return@mapNotNull null
                 val targetClass = clazz.javaFacade().findClass(qualifiedName, clazz.project.axonScope())
                     ?: return@mapNotNull null
-                val modelMember = inspect(targetClass) ?: return@mapNotNull null
+                val modelMember = inspect(targetClass, clazz) ?: return@mapNotNull null
                 val routingKey = field.resolveAnnotationStringValue(AxonAnnotation.AGGREGATE_MEMBER, "routingKey")
                 val eventForwardingMode = field.resolveAnnotationClassValue(AxonAnnotation.AGGREGATE_MEMBER, "eventForwardingMode")
-                ModelChild(field.name, modelMember, isCollection, routingKey, eventForwardingMode)
+                ModelChild(field, field.name, modelMember, isCollection, routingKey, eventForwardingMode)
             }
 
         val annotatedField = clazz.fields.firstOrNull { it.isAnnotated(AxonAnnotation.ENTITY_ID) }
@@ -107,7 +114,7 @@ class AggregateStructureResolver(private val project: Project) {
         } else {
             null
         }
-        return Model(clazz.qualifiedName!!, routingKey, children)
+        return Model(clazz.qualifiedName!!, clazz, parent?.qualifiedName, entityIdPresent, children)
     }
 
     /**